refactor(cli): group oclif core modules
