+++ src/lib/cli/argv-normalizer.ts
