+++ src/lib/cli/oclif-metadata.test.ts
